OAT readout
DigitalOAT.js was done in a way that was overriding the base file, affecting every plane that uses that file.  This way changing the folder structure keeps it separate to this plane.

Added L:var into the JS to select between C and F. Can then be selectable option in hanger and statesaved if wanted.
Added interaction on the screen to toggle between C and F
L:G36XIP_TEMP_SELECT  1=F 0=C
Added autodimming to OAT gauge.
Changed font and color based off the 1 blurry picture I could find of this thing lol